// examples/omnicounter-solana/src/generated/omnicounter/types/SetRemoteParams.ts
/**
 * This code was GENERATED using the solita package.
 * Please DO NOT EDIT THIS FILE, instead rerun solita to update it or write a wrapper to add functionality.
 *
 * See: https://github.com/metaplex-foundation/solita
 */

import * as beet from '@metaplex-foundation/beet'
export type SetRemoteParams = {
    id: number
    dstEid: number
    remote: number[] /* size: 32 */
}

/**
 * @category userTypes
 * @category generated
 */
export const setRemoteParamsBeet = new beet.BeetArgsStruct<SetRemoteParams>(
    [
        ['id', beet.u8],
        ['dstEid', beet.u32],
        ['remote', beet.uniformFixedSizeArray(beet.u8, 32)],
    ],
    'SetRemoteParams'
)