// examples/omnicounter-solana/src/generated/omnicounter/types/MessagingFee.ts
/**
 * This code was GENERATED using the solita package.
 * Please DO NOT EDIT THIS FILE, instead rerun solita to update it or write a wrapper to add functionality.
 *
 * See: https://github.com/metaplex-foundation/solita
 */

import * as beet from '@metaplex-foundation/beet'
export type MessagingFee = {
    nativeFee: beet.bignum
    lzTokenFee: beet.bignum
}

/**
 * @category userTypes
 * @category generated
 */
export const messagingFeeBeet = new beet.BeetArgsStruct<MessagingFee>(
    [
        ['nativeFee', beet.u64],
        ['lzTokenFee', beet.u64],
    ],
    'MessagingFee'
)