// examples/omnicounter-solana/src/generated/omnicounter/types/QuoteParams.ts
/**
 * This code was GENERATED using the solita package.
 * Please DO NOT EDIT THIS FILE, instead rerun solita to update it or write a wrapper to add functionality.
 *
 * See: https://github.com/metaplex-foundation/solita
 */

import * as beet from '@metaplex-foundation/beet'
export type QuoteParams = {
    dstEid: number
    receiver: number[] /* size: 32 */
    msgType: number
    options: Uint8Array
    payInLzToken: boolean
}

/**
 * @category userTypes
 * @category generated
 */
export const quoteParamsBeet = new beet.FixableBeetArgsStruct<QuoteParams>(
    [
        ['dstEid', beet.u32],
        ['receiver', beet.uniformFixedSizeArray(beet.u8, 32)],
        ['msgType', beet.u8],
        ['options', beet.bytes],
        ['payInLzToken', beet.bool],
    ],
    'QuoteParams'
)